JsonParserクラスの作成、テスト
