+++ Yumemi-Weather/Yumemi-WeatherTests/JsonParserTests.swift
@@ -8,9 +8,9 @@
 import XCTest
 @testable import Yumemi_Weather
 
-class JsonParseTests: XCTestCase {
+class JsonParserTests: XCTestCase {
     
-    var jsonPerser: JsonParser = WeatherModelImpl()
+    let jsonPerser = JsonParser()
     
     func test_Jsonのエンコード() throws {
         let expectedResult = """
@@ -32,23 +32,5 @@ class JsonParseTests: XCTestCase {
         XCTAssertEqual(result.minTemp, expectedResult.minTemp)
         XCTAssertEqual(result.date, expectedResult.date)
         XCTAssertEqual(result.weather, expectedResult.weather)
-    }
-}
-
-class JsonParserMock: JsonParser {
-    func request<T>(from value: T) throws -> String where T : Encodable {
-        let encoder = JSONEncoder()
-        let encodedDate = try encoder.encode(value)
-        
-        return String(data: encodedDate, encoding: .utf8)!
-    }
-    
-    func decode<T>(from value: String) throws -> T where T : Decodable {
-        let jsonData: Data = value.data(using: .utf8)!
-        let jsonDecoder = JSONDecoder()
-        jsonDecoder.keyDecodingStrategy = .convertFromSnakeCase
-        let decodedData = try jsonDecoder.decode(T.self, from: jsonData)
-        
-        return decodedData
     }
 }
